oncall_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-oncall-env"
version = "0.1.0"
description = "OnCallOps: Production Incident Response Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "oncall_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["oncall_env", "oncall_env.server"]
package-dir = { "oncall_env" = ".", "oncall_env.server" = "server" }
